Add whitelisting
Create option to whitelist lookups with 'only_for'. This config item
contains a list of scope keys to match the given regexes against.

Example:

:mysql_json:
  :only_for:
    :fqdn:
      - '^node01\.example\.com$'
    :operatingsystem:
      - 'CentOS'

Will cause the backend to perform lookups only if the operating system
matches CentOS or the fqdn is node01.example.com. It wil perform the
lookup if any of the conditions match.