Allow lookup constraints per back-end
It's currently possible to override database connection settings per
back-end. Expanding this functionality towards the only_for
configuration option that is used to limit lookups to specific
hosts/domains/etc.
Decided to keep the original point in code where this constraint is
checked for, so there's a possibility to return early, and prevent
additional files being opened.